requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-ninja==1.4.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
sqlparse==0.5.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
